Migration of checkout page
